User Stories
The description is a narrative that encompasses what the user will do, what happens when
everything goes right, and what to do if something goes wrong. It should include a reference to
any external user story that might be involved. For example, if a story requires that a user is
logged in, you should refer the reader to the ‘user log in’ story rather than explaining in detail
how the user is authenticated.
The goal is to be able to use a story as a guide during development. It should contain enough
information that a developer can understand what is needed and how the story relates to other
stories. You’ll also use your collected stories when deciding what sort of architecture and
infrastructure you will need. You can include diagrams if you think that they would be useful.